Fix handling of block comments after braces
Since the search_braces() function in indent.c loads _ALL_ comments
immediately surrounding braces into the save_com buffer, without
copying the indentation of the first line of the comment.

The code which copies the comment prepends it with a newline and a
space. The code in current_column() in code_io.c will only get the
newline and the space, returning a position of 2.

The code in print_comment() subtracts two from this number, printing the
whole thing verbatim as though it started in column 1.

This commit alters the print_comment() code in comments.c to check
whether or not the save buffer is being used, and if so, to use the
starting column set within the save buffer instead of calling
current_column() and subtracting 2, thus resulting in a properly
positioned comment.