-npcs: Fix recognition of function pointers
Indent reads a function pointer with more than one level of indirection
such as char **(*x) as 'char *' [decl]  '*' [unary_op] '(' [lparen] which
was preventing it from being detected as a function pointer since the code
for that was expecting only 'decl' before the lparen.

This patch fixes the detection code to account for this fact, since it
seems that there's other functionality relying on it.